Revert "Arrumado build no bohr."
