--- api/core/.env
@@ -1,1 +1,0 @@
-JWT_SECRET= 123lucas321
