--- api/core/api/index.js
@@ -1,4 +1,0 @@
-const express = require('@vendia/serverless-express')
-require("dotenv").config();
-const app = require('./routes')
-module.exports = express({ app })
