--- api/core/api/routes/index.js
@@ -1,12 +1,0 @@
-const bodyParser = require('body-parser')
-const routes = require('../../routes')
-const cors = require('cors')
-const authMiddleware = require('../middlewares/auth.middleware')
-
-module.exports = app => {
-    app.use(
-        bodyParser.json(),
-        cors(),
-        routes
-    )
-}
