--- api/core/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "orm-sequelize",
-  "version": "1.0.0",
-  "description": "",
-  "main": "./api/index.js",
-  "scripts": {
-    "dev": "nodemon ./api/index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@vendia/serverless-express": "^4.10.1",
-    "bcrypt": "^5.1.0",
-    "body-parser": "^1.19.0",
-    "cors": "^2.8.5",
-    "dotenv": "^16.0.3",
-    "express": "^4.18.2",
-    "jsonwebtoken": "^9.0.0",
-    "mysql2": "^2.1.0",
-    "path": "^0.12.7",
-    "sequelize": "^5.21.7",
-    "sequelize-cli": "^5.5.1"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.3"
-  }
-}
