+++ backend/.env
@@ -1,0 +1,1 @@
+JWT_SECRET= 123lucas321
